diabetes/requirements.txt
pandas==1.2.4
streamlit==1.10.0
numpy==1.23.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
